// docs/evaluations/current-ashiba-competitive-benchmark-v3/secondary-candidate-snapshots/AF-V-P-r1/corrected-final/src/prisma/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Ticket: {
      readonly id: CodecTypes['pg/int8@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly status: CodecTypes['pg/text@1']['output'];
      readonly assignee: CodecTypes['pg/text@1']['output'] | null;
      readonly priority: CodecTypes['pg/int4@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly metadata: CodecTypes['pg/json@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Ticket: {
      readonly id: CodecTypes['pg/int8@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
      readonly status: CodecTypes['pg/text@1']['input'];
      readonly assignee: CodecTypes['pg/text@1']['input'] | null;
      readonly priority: CodecTypes['pg/int4@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly metadata: CodecTypes['pg/json@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly tickets: {
      readonly assignee: CodecTypes['pg/text@1']['output'] | null;
      readonly created_at: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/int8@1']['output'];
      readonly metadata: CodecTypes['pg/json@1']['output'];
      readonly priority: CodecTypes['pg/int4@1']['output'];
      readonly status: CodecTypes['pg/text@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly tickets: {
      readonly assignee: CodecTypes['pg/text@1']['input'] | null;
      readonly created_at: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly id: CodecTypes['pg/int8@1']['input'];
      readonly metadata: CodecTypes['pg/json@1']['input'];
      readonly priority: CodecTypes['pg/int4@1']['input'];
      readonly status: CodecTypes['pg/text@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
    };
  };
};
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;